errors refinements
